requirements.txt
Flask==1.1.1
sklearn==0.0
gunicorn==19.9.0
python-dotenv==0.10.3


